Refactor Lambda PDF conversion into dedicated modules
